requirements.txt
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-cors-headers==3.11.0
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
firebase-admin==5.2.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-status==1.45.0
httplib2==0.20.4
idna==3.3
MarkupPy==1.14
msgpack==1.0.3
odfpy==1.4.1
openpyxl==3.0.9
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
